opencv-surveillance/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Environment variables
python-dotenv>=1.0.0

# Computer Vision & Face Recognition
opencv-contrib-python>=4.8.1.78  # Includes both opencv-python and contrib modules
numpy>=1.24.0
face_recognition>=1.3.0
pillow>=10.1.0

# dlib (required for face_recognition)
dlib>=19.24.0

# Email notifications
aiosmtplib>=2.0.0
email-validator>=2.1.0

# SMS notifications (Twilio)
twilio>=8.10.0

# Push notifications (Firebase - optional)
firebase-admin>=6.3.0

# Async HTTP for webhooks
aiohttp>=3.9.0

# Utilities
# Utilities
python-dateutil>=2.8.2

# Smart Home Integrations (Phase 4)
paho-mqtt>=1.6.1  # For Home Assistant MQTT
HAP-python[QRCode]>=4.9.0  # For HomeKit integration
google-auth>=2.25.0  # For Google Nest/Home
google-auth-oauthlib>=1.1.0
google-cloud-pubsub>=2.18.0  # For Google events

# Cloud Storage (Phase 5)
boto3>=1.34.0  # AWS S3
google-cloud-storage>=2.14.0  # Google Cloud Storage
azure-storage-blob>=12.19.0  # Azure Blob Storage

# Mobile Backend Support
websockets>=12.0  # For real-time updates to mobile

# PostgreSQL (Phase 6 - Production Database) - FREE & OPEN SOURCE
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Redis (Phase 6 - Caching) - FREE & OPEN SOURCE
redis>=5.0.1  # Redis client
hiredis>=2.2.3  # C parser for faster Redis operations

# Additional Security (Phase 6) - FREE & OPEN SOURCE
cryptography>=41.0.7  # Cryptographic operations
pyjwt>=2.8.0  # JSON Web Token handling
